+++ src/Trarizon.Library.Collections/HISTORY.md
@@ -5,4 +5,5 @@
 - Reimplement `PrefixTree(Dictionary)` with node, the original implementation is renamed to `ContiguousPrefixTree(Dicitionary)` and marked as experimental
 - Add `TraAlgorithm.LevenshteinDistance`
 - Add `IEnumerable.Intersperse`, `Interleave`, `RepeatInterleave`
+- Add `TraCollection.Singleton`
 - Remove `IEnumerable.LookAhead`